Start using cobra and viper
